requirements.txt
streamlit
pandas
scikit-learn
statsmodels
matplotlib
google-generativeai
Pillow
dotenv